FAG-2716 Switching library from Redis to MongoDB and working on unimplemented methods

+++ src/config/mod.rs
@@ -31,7 +31,7 @@ pub struct ServerConfig {
 #[derive(Clone, Deserialize)]
 pub struct MongoConfig {
     pub host: String,
-    pub port: Option<u32>,
+    pub port: Option<u16>,
     pub db_name: String,
     pub user: Option<String>,
     pub password: Option<String>,
